+++ lib/hexagon.dart
@@ -10,5 +10,6 @@ export 'src/grid/hexagon_grid.dart';
 export 'src/grid/hexagon_offset_grid.dart';
 export 'src/hexagon_clipper.dart';
 export 'src/hexagon_painter.dart';
+export 'src/hexagon_path_builder.dart';
 export 'src/hexagon_type.dart';
 export 'src/hexagon_widget.dart';